Refactored lifetime checker
